wheatley-rtc-pipeline/requirements.txt
livekit-agents>=0.12.1
livekit-plugins-openai>=0.10.9
livekit-plugins-silero>=0.7.4
livekit-plugins-elevenlabs>=0.7.10
livekit-api
python-dotenv~=1.0
